added userservice, activity diagram, sequence diagram, updated account controller and deleted myprofile viewmodel
